refactor(backend): reorganize the backend rust file. split the each feature into serveral file by template.
