+++ src-tauri/Cargo.toml
@@ -22,7 +22,7 @@ reqwest = { version = "0.13", features = ["cookies", "json"] }
 rusqlite = { version = "0.32", features = ["bundled"] }
 qrcode = { version = "0.14", default-features = false, features = ["svg"] }
 serde = { version = "1", features = ["derive"] }
-serde_json = "1"
+serde_json = { version = "1", features = ["preserve_order"] }
 tauri = { version = "2", features = [] }
 tauri-plugin-dialog = "2"
 tauri-plugin-opener = "2"
