+++ src-tauri/src/features/media/commands.rs
@@ -1,6 +1,5 @@
 //! media（FFmpeg）的 tauri command 薄壳。
-//! 目录/多选展开与 ffprobe 探测在此完成（IO 边界），参数决策全在 adapter 纯函数。
-//! inspect_media / expand_media_inputs / ffmpeg_encoders 是查询（§4.1），沿用 lib.rs 既有实现。
+//! 目录/多选展开与 ffprobe 探测归 query，参数决策归 adapter/policy。
 
 use serde::Serialize;
 use std::path::PathBuf;
@@ -8,7 +7,9 @@ use tauri::{AppHandle, State};
 use tokio::sync::OnceCell;
 
 use super::adapter::{self, MediaCtx, PrProbe};
+use super::query;
 use crate::core::adapter::{preview_result, PreviewResult};
+use crate::core::deps::{command_path, ffmpeg_encoders, resolve_tool, ToolName};
 use crate::core::task::types::{CwdPolicy, Feature, TaskIntent};
 use crate::core::task::{TaskHub, TaskSpec};
 
@@ -29,9 +30,7 @@ static ENCODER_FALLBACK: OnceCell<Option<String>> = OnceCell::const_new();
 async fn media_ctx(app: &AppHandle) -> MediaCtx {
     let fallback = ENCODER_FALLBACK
         .get_or_init(|| async {
-            let encoders = crate::ffmpeg_encoders(app.clone())
-                .await
-                .unwrap_or_default();
+            let encoders = ffmpeg_encoders(app.clone()).await.unwrap_or_default();
             FALLBACK_PREFERENCE
                 .iter()
                 .find(|name| encoders.iter().any(|e| e == *name))
@@ -64,6 +63,7 @@ fn ffmpeg_spec(
     let cwd = match plan.cwd {
         CwdPolicy::Inherit => None,
         CwdPolicy::ExeDir => tool_path.parent().map(|p| p.to_path_buf()),
+        CwdPolicy::Explicit(dir) => Some(std::path::PathBuf::from(dir)),
     };
     TaskSpec {
         feature: Feature::Media,
@@ -75,10 +75,12 @@ fn ffmpeg_spec(
         argv: plan.argv,
         argv_redacted: plan.argv_redacted,
         cwd,
-        env_path: Some(crate::command_path()),
+        output_paths: plan.output_paths,
+        env_path: Some(command_path()),
         intent, // media 无敏感字段，intent 无需 sanitize
         parser: None,
         on_failure: None,
+        cleanup_dir: None,
     }
 }
 
@@ -94,7 +96,7 @@ pub async fn media_submit(
         // 专家模式：argv 原文单任务提交
         let ctx = media_ctx(&app).await;
         let plan = adapter::plan(&intent, &ctx).map_err(|e| e.to_string())?;
-        let (tool_path, _) = crate::resolve_tool(&app, &crate::ToolName::Ffmpeg)
+        let (tool_path, _) = resolve_tool(&app, &ToolName::Ffmpeg)
             .ok_or_else(|| "未找到 FFmpeg，请先在依赖页安装".to_string())?;
         let id = hub.submit(ffmpeg_spec(plan, tool_path, intent.clone()));
         return Ok(BatchSubmitResult { task_ids: vec![id] });
@@ -104,26 +106,30 @@ pub async fn media_submit(
         .get("operation")
         .and_then(|v| v.as_str())
         .is_some_and(|op| op == "subtitle-extract");
-    let expanded = crate::expand_media_inputs(inputs, Some(include_subtitles))?;
+    let expanded = query::expand_media_inputs(inputs, Some(include_subtitles))?;
     if expanded.is_empty() {
         return Err("没有可处理的媒体文件".into());
     }
 
     let ctx = media_ctx(&app).await;
-    let (tool_path, _) = crate::resolve_tool(&app, &crate::ToolName::Ffmpeg)
+    let (tool_path, _) = resolve_tool(&app, &ToolName::Ffmpeg)
         .ok_or_else(|| "未找到 FFmpeg，请先在依赖页安装".to_string())?;
 
-    let mut task_ids = Vec::new();
-    for input in expanded {
-        // 每文件一个意图：重跑语义精确到单文件
-        let mut file_data = data.clone();
-        if let Some(map) = file_data.as_object_mut() {
-            map.insert("input".into(), serde_json::Value::String(input.clone()));
-        }
-        let file_intent = TaskIntent::Form(file_data);
-        let plan = adapter::plan(&file_intent, &ctx).map_err(|e| e.to_string())?;
-        task_ids.push(hub.submit(ffmpeg_spec(plan, tool_path.clone(), file_intent)));
-    }
+    // 先为全部文件完成 plan，确保第一个任务入队后不再有可失败的准备步骤。
+    let specs = expanded
+        .into_iter()
+        .map(|input| {
+            // 每文件一个意图：重跑语义精确到单文件
+            let mut file_data = data.clone();
+            if let Some(map) = file_data.as_object_mut() {
+                map.insert("input".into(), serde_json::Value::String(input));
+            }
+            let file_intent = TaskIntent::Form(file_data);
+            let plan = adapter::plan(&file_intent, &ctx).map_err(|e| e.to_string())?;
+            Ok(ffmpeg_spec(plan, tool_path.clone(), file_intent))
+        })
+        .collect::<Result<Vec<_>, String>>()?;
+    let task_ids = specs.into_iter().map(|spec| hub.submit(spec)).collect();
     Ok(BatchSubmitResult { task_ids })
 }
 
@@ -132,41 +138,68 @@ pub async fn media_submit(
 pub async fn media_pr_submit(
     app: AppHandle,
     hub: State<'_, TaskHub>,
-    input: String,
+    inputs: Vec<String>,
     output_directory: Option<String>,
 ) -> Result<BatchSubmitResult, String> {
-    let (tool_path, _) = crate::resolve_tool(&app, &crate::ToolName::Ffmpeg)
+    let (tool_path, _) = resolve_tool(&app, &ToolName::Ffmpeg)
         .ok_or_else(|| "未找到 FFmpeg，请先在依赖页安装".to_string())?;
-    let (ffprobe, _) = crate::resolve_tool(&app, &crate::ToolName::Ffprobe)
+    let (ffprobe, _) = resolve_tool(&app, &ToolName::Ffprobe)
         .ok_or_else(|| "未找到 ffprobe，请先在依赖页安装".to_string())?;
 
-    let input_path = PathBuf::from(&input);
-    let inputs = if input_path.is_dir() {
-        crate::media_files_in(&input_path)?
-    } else if input_path.is_file() {
-        vec![input_path]
-    } else {
-        return Err("输入文件或目录不存在".into());
-    };
+    let mut expanded = Vec::new();
+    for input in inputs {
+        let input_path = PathBuf::from(input);
+        if input_path.is_dir() {
+            expanded.extend(query::media_files_in(&input_path)?);
+        } else if input_path.is_file() {
+            expanded.push(input_path);
+        } else {
+            return Err("输入文件或目录不存在".into());
+        }
+    }
+    expanded.sort();
+    expanded.dedup();
+    if expanded.is_empty() {
+        return Err("没有可处理的媒体文件".into());
+    }
 
-    let mut task_ids = Vec::new();
-    for path in inputs {
-        let (video, audio, subtitles) = crate::probe_streams(&ffprobe, &path).await?;
+    // 探测与 plan 全部完成后再准备目录；任一步失败都不会产生部分入队。
+    let mut prepared = Vec::with_capacity(expanded.len());
+    for path in expanded {
+        let (video, audio, subtitles) = query::probe_streams(&ffprobe, &path).await?;
         let probe = PrProbe {
             video,
             audio,
             subtitles,
         };
         let (plan, output) = adapter::pr_plan(&path, &probe, output_directory.as_deref())?;
-        if let Some(parent) = output.parent() {
-            std::fs::create_dir_all(parent).map_err(|e| e.to_string())?;
-        }
         let intent = TaskIntent::Form(serde_json::json!({
             "prCompatible": true,
             "input": path.to_string_lossy(),
-            "outputDirectory": output_directory,
+            "outputDirectory": output_directory.clone(),
         }));
-        task_ids.push(hub.submit(ffmpeg_spec(plan, tool_path.clone(), intent)));
+        prepared.push((plan, output, intent));
     }
+
+    for (_, output, _) in &prepared {
+        if let Some(parent) = output.parent() {
+            std::fs::create_dir_all(parent).map_err(|e| e.to_string())?;
+        }
+    }
+
+    let specs = prepared
+        .into_iter()
+        .map(|(plan, _, intent)| ffmpeg_spec(plan, tool_path.clone(), intent));
+    let task_ids = specs.map(|spec| hub.submit(spec)).collect();
     Ok(BatchSubmitResult { task_ids })
+}
+
+/// 输入预扫描：目录递归展开为媒体文件，供前端文件列表展示。
+/// 复用提交时的 expand_media_inputs，保证与实际执行语义一致。
+#[tauri::command]
+pub fn media_scan_inputs(
+    inputs: Vec<String>,
+    include_subtitles: bool,
+) -> Result<Vec<String>, String> {
+    query::expand_media_inputs(inputs, Some(include_subtitles))
 }
